Removed old project files
It is better to decouple storage subsystem (database)
from the app logic. Django is not the most suitable framework
in this regard. Besides its current code is syncronous (WSGI).

The alternative is to use FastAPI as a web-framework and a separate
framework to handle the persistence of the data (sqlalchemy, for example).

Signed-off-by: Nikita Vakula <programmistov.programmist@gmail.com>